Move request and response to http.go
